fix: map wildcard CLI hosts and keep stderr log follow
Route CLI API clients through loopback when server.host is a wildcard bind
so Host allowlist accepts daemon status/config calls. Always open a
secondary stderr=1 log stream so later stderr is not dropped after default
follow switches to stdout. Split browser guard tests under the 300-line gate.

Generated-By: looper 0.0.0-dev (runner=fixer, agent=grok-build)
