+++ internal/cliapp/daemon_runtime.go
@@ -862,7 +862,10 @@ func (r *commandRuntime) apiClientFromLoaded(loaded config.LoadedFileConfig) *Da
 		baseURL = strings.TrimSpace(*loaded.Config.Server.BaseURL)
 	} else {
 		// Bracket IPv6 literals (net.JoinHostPort) so dial works for hosts like ::1.
-		baseURL = dashboardHTTPBaseURL(loaded.Config.Server.Host, loaded.Config.Server.Port)
+		// Map wildcard binds (0.0.0.0 / ::) to loopback so request Host matches the
+		// browser Host guard allowlist (same mapping as dashboardAPIClient).
+		host := browserHostForDashboard(loaded.Config.Server.Host)
+		baseURL = dashboardHTTPBaseURL(host, loaded.Config.Server.Port)
 	}
 
 	return r.newAPIClientForLoaded(loaded, baseURL)
@@ -870,8 +873,10 @@ func (r *commandRuntime) apiClientFromLoaded(loaded config.LoadedFileConfig) *Da
 
 func (r *commandRuntime) localAPIClientFromLoaded(loaded config.LoadedFileConfig) *DaemonAPIClient {
 	// Reuse dashboardHTTPBaseURL so daemonStart readiness polling brackets IPv6
-	// hosts the same way as `looper dashboard` probe/mint clients.
-	baseURL := dashboardHTTPBaseURL(loaded.Config.Server.Host, loaded.Config.Server.Port)
+	// hosts the same way as `looper dashboard` probe/mint clients. Map wildcard
+	// binds to loopback so Host is allowlisted by the browser guard.
+	host := browserHostForDashboard(loaded.Config.Server.Host)
+	baseURL := dashboardHTTPBaseURL(host, loaded.Config.Server.Port)
 	return r.newAPIClientForLoaded(loaded, baseURL)
 }
